introduce userSettingsWrapper and delete redundant  accountProfile component
